% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_multiple_passwords.R
\name{generate_multiple_passwords}
\alias{generate_multiple_passwords}
\title{Generate multiple passwords}
\usage{
generate_multiple_passwords(
  n = 10,
  length = 12,
  use_upper = TRUE,
  use_lower = TRUE,
  use_numbers = TRUE,
  use_special = TRUE
)
}
\arguments{
\item{n}{Number of passwords to generate.}

\item{length}{Length of each password.}

\item{use_upper}{Logical, use uppercase letters.}

\item{use_lower}{Logical, use lowercase letters.}

\item{use_numbers}{Logical, use numbers.}

\item{use_special}{Logical, use special characters.}
}
\value{
A vector of randomly generated passwords.
}
\description{
Generate multiple passwords
}
\examples{
# Generate 5 passwords, each 16 characters long with all types of characters
generate_multiple_passwords(5, 16, TRUE, TRUE, TRUE, TRUE)

# Generate 3 passwords, each 12 characters long with only lowercase and numbers
generate_multiple_passwords(3, 12, use_upper = FALSE, use_lower = TRUE, use_numbers = TRUE, use_special = FALSE)
}
