% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_password.R
\name{generate_password}
\alias{generate_password}
\title{Generate a single password}
\usage{
generate_password(
  length = 12,
  use_upper = TRUE,
  use_lower = TRUE,
  use_numbers = TRUE,
  use_special = TRUE
)
}
\arguments{
\item{length}{Length of the password.}

\item{use_upper}{Logical, use uppercase letters.}

\item{use_lower}{Logical, use lowercase letters.}

\item{use_numbers}{Logical, use numbers.}

\item{use_special}{Logical, use special characters.}
}
\value{
A randomly generated password.
}
\description{
Generate a single password
}
\examples{
# Generate a 16-character password with all types of characters
generate_password(16, TRUE, TRUE, TRUE, TRUE)

# Generate a 12-character password with only lowercase and numbers
generate_password(12, use_upper = FALSE, use_lower = TRUE, use_numbers = TRUE, use_special = FALSE)
}
